rsamara-package {recomeristem}	R Documentation

A short title line describing what the package does

Description

A more detailed description of what the package does. A length of about one to five lines is recommended.

Details

This section should provide a more detailed overview of how to use the package, including the most important functions.

Author(s)

Eric Ramat, email optional.

Maintainer: Eric Ramat <eric.ramat@gmail.com>

References

This optional section can contain literature or other references for background information.

See Also

Optional links to other man pages

Examples

  ## Not run: 
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  
## End(Not run)

[Package recomeristem version 1.0 Index]